+++ platform-lookup-service/src/test/java/com/emaratech/platform/lookupsvc/test/CityTest.java
@@ -2,15 +2,13 @@ package com.emaratech.platform.lookupsvc.test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.math.BigDecimal;
-
 import org.junit.jupiter.api.Test;
 
 import com.emaratech.platform.lookupsvc.model.City;
 
 public class CityTest {
 
-    private BigDecimal id = BigDecimal.valueOf(1L);
+    private Long id = 1L;
 
     /**
      * Tests the getter for City.
